@@ -1,42 +1,44 @@
 {
   "name": "shane-sfdx-plugins",
   "description": "sfdx plugins by Shane McLaughlin",
-  "version": "0.19.0",
+  "version": "0.20.0",
   "author": "@mshanemc",
   "bugs": "https://github.com/mshanemc/shane-sfdx-plugins/issues",
   "dependencies": {
-    "@oclif/command": "^1.4.7",
-    "@oclif/config": "^1.3.62",
-    "@oclif/errors": "^1.0.3",
-    "@salesforce/command": "0.0.10",
-    "@types/lodash": "^4.14.110",
-    "chalk": "^2.3.2",
+    "@oclif/command": "^1.4.34",
+    "@oclif/config": "^1.6.33",
+    "@oclif/errors": "^1.1.2",
+    "@oclif/parser": "^3.5.3",
+    "@salesforce/command": "0.1.6",
+    "chalk": "^2.4.1",
     "fs-extra": "^5.0.0",
     "js2xmlparser": "^3.0.0",
     "lodash": "^4.17.10",
     "npm": "^5.10.0",
-    "request": "^2.85.0",
+    "request": "^2.87.0",
     "request-promise-native": "^1.0.5",
     "xml2js": "^0.4.19"
   },
   "devDependencies": {
-    "@oclif/dev-cli": "^1.13.0",
-    "@oclif/plugin-help": "^1.2.2",
-    "@oclif/test": "^1.0.1",
-    "@oclif/tslint": "^1.0.2",
-    "@salesforce/dev-config": "^1.0.4",
-    "@types/chai": "^4.1.2",
-    "@types/fs-extra": "^5.0.1",
-    "@types/mocha": "^5.0.0",
-    "@types/node": "^9.6.2",
-    "@types/request-promise-native": "^1.0.14",
+    "@oclif/dev-cli": "^1.15.4",
+    "@oclif/plugin-help": "^1.2.11",
+    "@oclif/test": "^1.1.0",
+    "@oclif/tslint": "^1.1.2",
+    "@salesforce/dev-config": "^1.1.2",
+    "@types/chai": "^4.1.4",
+    "@types/fs-extra": "^5.0.3",
+    "@types/jsforce": "^1.8.13",
+    "@types/lodash": "^4.14.110",
+    "@types/mocha": "^5.2.4",
+    "@types/node": "^9.6.23",
+    "@types/request-promise-native": "^1.0.15",
     "chai": "^4.1.2",
     "globby": "^8.0.1",
-    "mocha": "^5.0.5",
+    "mocha": "^5.2.0",
     "sinon": "^4.5.0",
     "ts-node": "5",
-    "tslint": "^5.9.1",
-    "typescript": "^2.8.1"
+    "tslint": "^5.10.0",
+    "typescript": "^2.9.2"
   },
   "engines": {
     "node": ">=8.0.0"
