@@ -1,14 +1,11 @@
-import { flags } from '@oclif/command';
-import { join } from 'path';
 import { SfdxCommand, core } from '@salesforce/command';
-import fs = require('fs-extra');
 import util = require('util');
 import child_process = require('child_process');
+import ucc = require('../../../../shared/unzipConvertClean');
 
 const exec = util.promisify(child_process.exec);
 
 const tmpDir = 'mdapiout';
-const pkgDir = 'pkgDirTemp';
 
 export default class Get extends SfdxCommand {
 
@@ -26,8 +23,8 @@ export default class Get extends SfdxCommand {
   protected static requiresUsername = true;
 
   protected static flagsConfig = {
-    packageName: flags.string({ required: true, char: 'p', description: 'the name of the package you want to retrieve	' }),
-    target: flags.string({ char: 't', default: 'force-app', description: 'where to convert the result to...defaults to force-app' })
+    packagename: {type: 'string', required: true, char: 'p', description: 'the name of the package you want to retrieve	' },
+    target: {type: 'string', char: 't', default: 'force-app', description: 'where to convert the result to...defaults to force-app' }
   };
 
   protected static requiresProject = true;
@@ -35,18 +32,8 @@ export default class Get extends SfdxCommand {
   // tslint:disable-next-line:no-any
   public async run(): Promise<any> {
 
-    process.stdout.write('starting package retrieval...');
+    const retrieveCommand = `sfdx force:mdapi:retrieve -s -p "${this.flags.packagename}" -u ${this.org.getUsername()}  -r ./${tmpDir} -w 30`;
+    await ucc.retrieveUnzipConvertClean(tmpDir, retrieveCommand, this.flags.target);
 
-    await exec(`sfdx force:mdapi:retrieve -s -p "${this.flags.packageName}" -u ${this.org.getUsername()}  -r ./${tmpDir} -w 30`);
-    process.stdout.write('Package Retrieved.  Unzipping...');
-
-    await exec(`unzip -o ./${tmpDir}/unpackaged.zip -d ./${tmpDir}`);
-    process.stdout.write('Package Unzipped.  Converting...');
-
-    await exec(`sfdx force:mdapi:convert -r ./${tmpDir} -d ${this.flags.target}`);
-    process.stdout.write('Package Converted.  Cleaning up...');
-
-    await exec(`rm -rf ./${tmpDir}`);
-    this.ux.log('Done!');
   }
 }
