@@ -6,7 +6,7 @@ import cli from 'cli-ux';
 import jsToXml = require('js2xmlparser');
 
 import chalk from 'chalk';
-import { getExisting } from '../../../shared/getExisting';
+import { getExisting, fixExistingDollarSign } from '../../../shared/getExisting';
 import * as options from '../../../shared/js2xmlStandardOptions';
 
 export default class FATUpdate extends SfdxCommand {
@@ -30,8 +30,8 @@ export default class FATUpdate extends SfdxCommand {
 
   protected static flagsConfig = {
     object: {char: 'o', description: 'object to manage the policy for', type: 'string', required: true},
-    archiveAfterMonths: { char: 'm', description: 'archive after this number of months', type: 'number', default: 18 },
-    archiveRetentionYears: { char: 'y', default: 10, description: 'Archive for this many years', type: 'number' },
+    archiveaftermonths: { char: 'm', description: 'archive after this number of months', type: 'number', default: 18 },
+    archiveretentionyears: { char: 'y', default: 10, description: 'Archive for this many years', type: 'number' },
     description: {description: 'optional friendly description for the policy', type: 'string'},
     directory: { char: 'd', default: 'force-app/main/default', description: 'Where is all this metadata? defaults to force-app/main/default', type: 'string' }
   };
@@ -53,27 +53,22 @@ export default class FATUpdate extends SfdxCommand {
       return;
     }
 
-    const existing = await getExisting(targetFilename, 'CustomObject');
+    let existing = await getExisting(targetFilename, 'CustomObject');
     existing.enableHistory = true;
 
     existing.historyRetentionPolicy = existing.historyRetentionPolicy || {};
 
-    if (this.flags.archiveAfterMonths) {
-      existing.historyRetentionPolicy.archiveAfterMonths = this.flags.archiveAfterMonths;
+    if (this.flags.archiveaftermonths) {
+      existing.historyRetentionPolicy.archiveAfterMonths = this.flags.archiveaftermonths;
     }
-    if (this.flags.archiveRetentionYears) {
-      existing.historyRetentionPolicy.archiveRetentionYears = this.flags.archiveRetentionYears;
+    if (this.flags.archiveretentionyears) {
+      existing.historyRetentionPolicy.archiveRetentionYears = this.flags.archiveretentionyears;
     }
     if (this.flags.description) {
       existing.historyRetentionPolicy.description = this.flags.description;
     }
 
-    // correct @ => $ issue
-    if (existing['$']) {
-      const temp = existing['$'];
-      delete existing['$'];
-      existing['@'] = temp;
-    }
+    existing = await fixExistingDollarSign(existing);
 
     // convert to xml and write out the file
     const xml = jsToXml.parse('CustomObject', existing, options.js2xmlStandardOptions);
