@@ -1,14 +1,8 @@
-import { flags } from '@oclif/command';
-import { join } from 'path';
 import { SfdxCommand, core } from '@salesforce/command';
 import request = require('request-promise-native');
 import localFile2CV = require('../../../shared/localFile2CV');
 import userIdLookup = require('../../../shared/userIdLookup');
-
 import chalk from 'chalk';
-
-core.Messages.importMessagesDirectory(join(__dirname, '..', '..', '..'));
-// const messages = core.Messages.loadMessages('shane-sfdx-plugins', 'org');
 
 export default class Photo extends SfdxCommand {
 
@@ -27,10 +21,10 @@ export default class Photo extends SfdxCommand {
   ];
 
   protected static flagsConfig = {
-    firstName: flags.string({ char: 'g', description: 'first (given) name of the user--keeping -f for file for consistency' }),
-    lastName: flags.string({ char: 'l', required: true, description: 'last name of the user' }),
-    file: flags.string({ char: 'f', description: 'local path of the photo to use' }),
-    banner: flags.string({ char: 'b', description: 'local path of the chatter banner photo to use' })
+    firstname: { type: 'string',  char: 'g', description: 'first (given) name of the user--keeping -f for file for consistency'},
+    lastname: { type: 'string',  char: 'l', required: true, description: 'last name of the user' },
+    file: { type: 'string',  char: 'f', description: 'local path of the photo to use' },
+    banner: { type: 'string',  char: 'b', description: 'local path of the chatter banner photo to use' }
   };
 
   // Comment this out if your command does not require an org username
@@ -63,7 +57,7 @@ export default class Photo extends SfdxCommand {
     let user;
 
     try {
-      user = await userIdLookup.getUserId(conn, this.flags.lastName, this.flags.firstName);
+      user = await userIdLookup.getUserId(conn, this.flags.lastname, this.flags.firstname);
     } catch (e) {
       this.ux.error(chalk.red(e));
       return {
