@@ -16,12 +16,12 @@ const api = 'Hello__e';
 const label = 'Hello';
 const plural = 'Hellos';
 
-before(async () => {
-  await exec(`rm -rf ${testProjectName}`);
-  await exec(`sfdx force:project:create -n ${testProjectName}`);
-});
-
 describe('shane:object:create (platform event flavor)', () => {
+
+  before(async function() {
+    await exec(`rm -rf ${testProjectName}`);
+    await exec(`sfdx force:project:create -n ${testProjectName}`);
+  });
 
   it('creates a platform event with all params supplied', async () => {
 
@@ -39,7 +39,7 @@ describe('shane:object:create (platform event flavor)', () => {
     expect(parsed.CustomObject.label).to.equal(label);
     expect(parsed.CustomObject.pluralLabel).to.equal(plural);
     expect(parsed.CustomObject.eventType).to.equal('StandardVolume');
-  }).timeout(60000);
+  });
 
   it('creates a Number field (18,0) on the Event', async () => {
 
@@ -60,7 +60,7 @@ describe('shane:object:create (platform event flavor)', () => {
 
     const parsedObj = await testutils.getParsedXML(`${testProjectName}/force-app/main/default/objects/${api}/${api}.object-meta.xml`);
 
-  }).timeout(60000);
+  });
 
   it('creates a Text field on the Event', async () => {
 
@@ -80,7 +80,7 @@ describe('shane:object:create (platform event flavor)', () => {
 
     const parsedObj = await testutils.getParsedXML(`${testProjectName}/force-app/main/default/objects/${api}/${api}.object-meta.xml`);
 
-  }).timeout(60000);
+  });
 
   it('creates a required text field on the Event', async () => {
 
@@ -101,7 +101,7 @@ describe('shane:object:create (platform event flavor)', () => {
 
     const parsedObj = await testutils.getParsedXML(`${testProjectName}/force-app/main/default/objects/${api}/${api}.object-meta.xml`);
 
-  }).timeout(60000);
+  });
 
   it('creates a checkbox field on the Event', async () => {
 
@@ -121,7 +121,7 @@ describe('shane:object:create (platform event flavor)', () => {
 
     const parsedObj = await testutils.getParsedXML(`${testProjectName}/force-app/main/default/objects/${api}/${api}.object-meta.xml`);
 
-  }).timeout(60000);
+  });
 
   it('can build a permset', async () => {
     const permSetName = 'MyEventPerm';
@@ -149,7 +149,7 @@ describe('shane:object:create (platform event flavor)', () => {
 
     expect(parsed.PermissionSet.fieldPermissions).to.not.deep.include({ readable: 'true', editable: 'true', field: `${api}.Required_Text_Field__c` });
 
-  }).timeout(60000);
+  });
 
   it('deploys as valid code', async () => {
     if (process.env.LOCALONLY === 'true') {
@@ -158,10 +158,9 @@ describe('shane:object:create (platform event flavor)', () => {
       const deploySuccess = await testutils.itDeploys(testProjectName);
       expect(deploySuccess).to.be.true;
     }
-  }).timeout(60000);
+  });
 
-});
-
-after(async () => {
-  await exec(`rm -rf ${testProjectName}`);
+  after(async () => {
+    await exec(`rm -rf ${testProjectName}`);
+  });
 });
