@@ -12,12 +12,12 @@ import testutils = require('../../../helpers/testutils');
 const exec = util.promisify(child_process.exec);
 const testProjectName = 'testProject';
 
-before(async () => {
-  await exec(`rm -rf ${testProjectName}`);
-  await exec(`sfdx force:project:create -n ${testProjectName}`);
-});
-
 describe('shane:remotesite:create', () => {
+
+  before(async function() {
+    await exec(`rm -rf ${testProjectName}`);
+    await exec(`sfdx force:project:create -n ${testProjectName}`);
+  });
 
   it('creates a RemSite from url + name', async () => {
 
@@ -58,10 +58,9 @@ describe('shane:remotesite:create', () => {
       const deploySuccess = await testutils.itDeploys(testProjectName);
       expect(deploySuccess).to.be.true;
     }
-  }).timeout(60000);
+  });
 
-});
-
-after( async () => {
-  await exec(`rm -rf ${testProjectName}`);
+  after( async () => {
+    await exec(`rm -rf ${testProjectName}`);
+  });
 });
