@@ -1,3 +1,3 @@
 const path = require('path')
 process.env.TS_NODE_PROJECT = path.resolve('test/tsconfig.json')
-process.env.LOCALONLY = true;
+process.env.LOCALONLY = false;
